requirements.txt
multipledispatch==1.0.0
numpy==1.26.4
pillow==10.3.0
pygame==2.5.2
PyOpenGL==3.1.7
pyrr==0.10.3
